requirements.txt
alembic==1.5.8
click==7.1.2
ecdsa==0.17.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
